+++ predicators/ground_truth_models/playroom/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for playroom environment and variants."""
+
+from .nsrts import PlayroomGroundTruthNSRTFactory
+
+__all__ = ["PlayroomGroundTruthNSRTFactory"]
